+++ src/ArduinoJson/JsonObjectImpl.hpp
@@ -14,8 +14,10 @@ template <typename TStringRef>
 inline JsonArray &JsonObject::createNestedArray_impl(TStringRef key) {
   if (!_buffer) return JsonArray::invalid();
   JsonArray &array = _buffer->createArray();
-  set(key, array);
-  return array;
+  if (set(key, array))
+    return array;
+  else
+    return JsonArray::invalid();
 }
 
 template <typename TStringRef>